Fix GCC builds. Move a few templates from .hpp to .cpp files
